# dashboard/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Memory Observer Dashboard API
  description: |
    REST API for the Memory Observer Dashboard, providing access to workspaces, plans, agents, and events
    from the Project Memory MCP Server.
    
    ## Overview
    
    The Memory Observer Dashboard API allows you to:
    - View and manage MCP workspaces
    - Browse, create, and manage project plans
    - View and deploy agent templates
    - Subscribe to real-time MCP events
    - Search across all workspaces and plans
    
    ## Authentication
    
    Currently, the API runs locally and does not require authentication.
    Future versions may support authentication for remote access.
    
  version: 0.1.0
  contact:
    name: Project Memory MCP
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local Development Server
...
paths:
...
  /api/agents:
    get:
      tags:
        - Agents
      summary: List all agent templates
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Agents
      summary: Create a new agent template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent template details
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details with content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - Agents
      summary: Update agent template
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Updated markdown content
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Agents
      summary: Delete agent template
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/agents/{agentId}/deploy:
    post:
      tags:
        - Agents
      summary: Deploy agent to workspaces
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workspaceIds
              properties:
                workspaceIds:
                  type: array
                  items:
                    type: string
                  description: List of workspace IDs to deploy to
      responses:
        '200':
          description: Agent deployed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deployed:
                    type: array
                    items:
                      type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/agents/{agentId}/diff/{workspaceId}:
    get:
      tags:
        - Agents
      summary: Get diff between template and deployed version
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Diff result
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasChanges:
                    type: boolean
                  templateHash:
                    type: string
                  deployedHash:
                    type: string
                  diff:
                    type: array
                    items:
                      $ref: '#/components/schemas/DiffLine'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        example: ds_my-project-abc123
...
    AgentId:
      name: agentId
      in: path
      required: true
      description: Agent template ID (filename without .agent.md)
      schema:
        type: string
        example: coordinator
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request parameters
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          example: coordinator
        name:
          type: string
          example: Coordinator
        description:
          type: string
        lastModified:
          type: string
          format: date-time
        deployedCount:
          type: integer
        isTemplate:
          type: boolean
    AgentDetails:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            content:
              type: string
              description: Raw markdown content
            deployments:
              type: array
              items:
                type: object
                properties:
                  workspaceId:
                    type: string
                  workspaceName:
                    type: string
                  syncStatus:
                    type: string
                    enum: [synced, outdated, customized, missing]
                  hash:
                    type: string
    CreateAgentRequest:
      type: object
      required:
        - name
        - content
      properties:
        name:
          type: string
          description: Agent name (will be converted to filename)
        content:
          type: string
          description: Markdown content for the agent template
    DiffLine:
      type: object
      properties:
        type:
          type: string
          enum: [unchanged, added, removed]
        lineNumber:
          type: integer
        content:
          type: string
...